refactor(react-native): move app start logic to span factory
