--- packages/platforms/react-native/lib/create-app-start-span.ts
@@ -1,16 +1,0 @@
-import type { SpanInternal } from '@bugsnag/core-performance'
-import type { ReactNativeSpanFactory } from './span-factory'
-
-let appStartSpan: SpanInternal
-
-export function createAppStartSpan (spanFactory: ReactNativeSpanFactory, appStartTime: number) {
-  if (appStartSpan) {
-    return appStartSpan
-  }
-
-  appStartSpan = spanFactory.startSpan('[AppStart/ReactNativeInit]', { startTime: appStartTime, parentContext: null })
-  appStartSpan.setAttribute('bugsnag.span.category', 'app_start')
-  appStartSpan.setAttribute('bugsnag.app_start.type', 'ReactNativeInit')
-
-  return appStartSpan
-}
